@@ -19,7 +19,7 @@ repository: https://github.com/Digia-Technology-Private-Limited/digia_ui.git
 version: 1.6.1
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: ">=3.3.0 <4.0.0"
 
 # https://github.com/frencojobs/derry
 scripts:
@@ -72,6 +72,7 @@ dependencies:
   uuid: ^4.5.1
   video_player: ^2.10.0
   webview_flutter: ^4.13.0
+  youtube_player_flutter: ^9.1.3
   youtube_player_iframe: ^5.2.2
   table_calendar: ^3.0.9
   pinput: ^5.0.1
@@ -143,5 +144,5 @@ flutter:
   # see https://flutter.dev/custom-fonts/#from-packages
 
 platforms:
-   android:
-   ios:
+  android:
+  ios:
